split interactive tests
